% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openrbqm-package.R
\docType{package}
\name{openrbqm-package}
\alias{openrbqm}
\alias{openrbqm-package}
\title{openrbqm: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openrbqm.github.io/openrbqm/}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Laura Maxwell \email{lkmaxwell23@gmail.com}
}

}
\keyword{internal}
