% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{openrbqm_packages}
\alias{openrbqm_packages}
\title{List all packages imported by openrbqm}
\usage{
openrbqm_packages()
}
\description{
List all packages imported by openrbqm
}
\examples{
#openrbqm_packages()
}
